+++ orwynn/apprc/__init__.py
@@ -1,2 +1,2 @@
 from .apprc import AppRc
-from .parse import parse_apprc
+from .utils import AppRCUtils