+++ orwynn/helpers/constants.py
@@ -20,7 +20,7 @@ SUBCLASSABLES: list[BaseSubclassable] = [
 ENVIRONS: set[str] = {
     "ORWYNN_MODE",
     "ORWYNN_ROOT_DIR",
-    "ORWYNN_APPRC_PATH",
+    "ORWYNN_RC_PATH",
     "ORWYNN_IS_CATCH_LOGGING_ENABLED_IN_TESTS"
 }
 """